+++ src/systems.rs
@@ -2,6 +2,7 @@
 
 #[cfg(feature = "ui")]
 use crate::action_driver::ActionStateDriver;
+use crate::conflicting_inputs::TrackingInputType;
 use crate::{
     action_state::ActionState, clashing_inputs::ClashStrategy, input_map::InputMap,
     input_streams::InputStreams, plugin::ToggleActions, Actionlike,
@@ -22,11 +23,8 @@ use bevy::{
 };
 
 use crate::action_diff::{ActionDiff, ActionDiffEvent};
-
 #[cfg(feature = "ui")]
 use bevy::ui::Interaction;
-#[cfg(feature = "egui")]
-use bevy_egui::EguiContext;
 
 /// Advances actions timer.
 ///
@@ -72,51 +70,36 @@ pub fn update_action_state<A: Actionlike>(
     mut mouse_wheel: EventReader<MouseWheel>,
     mut mouse_motion: EventReader<MouseMotion>,
     clash_strategy: Res<ClashStrategy>,
-    #[cfg(all(feature = "ui", not(feature = "no_ui_priority")))] interactions: Query<&Interaction>,
-    #[cfg(feature = "egui")] mut maybe_egui: Query<(Entity, &'static mut EguiContext)>,
+    mut tracking_input: ResMut<TrackingInputType>,
     action_state: Option<ResMut<ActionState<A>>>,
     input_map: Option<Res<InputMap<A>>>,
     mut query: Query<(&mut ActionState<A>, &InputMap<A>)>,
 ) {
-    let gamepad_buttons = gamepad_buttons.into_inner();
-    let gamepad_button_axes = gamepad_button_axes.into_inner();
-    let gamepad_axes = gamepad_axes.into_inner();
-    let gamepads = gamepads.into_inner();
-    let keycodes = keycodes.map(|keycodes| keycodes.into_inner());
-    let mouse_buttons = mouse_buttons.map(|mouse_buttons| mouse_buttons.into_inner());
-
-    let mouse_wheel: Option<Vec<MouseWheel>> = Some(mouse_wheel.read().cloned().collect());
-    let mouse_motion: Vec<MouseMotion> = mouse_motion.read().cloned().collect();
-
-    // If use clicks on a button, do not apply them to the game state
-    #[cfg(all(feature = "ui", not(feature = "no_ui_priority")))]
-    let (mouse_buttons, mouse_wheel) = if interactions
-        .iter()
-        .any(|&interaction| interaction != Interaction::None)
-    {
-        (None, None)
-    } else {
-        (mouse_buttons, mouse_wheel)
-    };
-
-    // If egui wants to own inputs, don't also apply them to the game state
-    #[cfg(feature = "egui")]
-    let keycodes = maybe_egui
-        .iter_mut()
-        .all(|(_, mut ctx)| !ctx.get_mut().wants_keyboard_input())
-        .then_some(keycodes)
+    // Track these inputs only when the corresponding flags are enabled
+    let gamepad_buttons = tracking_input.gamepad.then(|| gamepad_buttons.into_inner());
+    let gamepad_button_axes = tracking_input
+        .gamepad
+        .then(|| gamepad_button_axes.into_inner());
+    let gamepad_axes = tracking_input.gamepad.then(|| gamepad_axes.into_inner());
+    let gamepads = tracking_input.gamepad.then(|| gamepads.into_inner());
+    let keycodes: Option<&ButtonInput<KeyCode>> = tracking_input
+        .keyboard
+        .then(|| keycodes.map(|keycodes| keycodes.into_inner()))
         .flatten();
+    let mouse_buttons: Option<&ButtonInput<MouseButton>> = tracking_input
+        .mouse
+        .then(|| mouse_buttons.map(|mouse_buttons| mouse_buttons.into_inner()))
+        .flatten();
+    let mouse_wheel: Option<Vec<MouseWheel>> = tracking_input
+        .mouse
+        .then(|| mouse_wheel.read().cloned().collect());
+    let mouse_motion: Vec<MouseMotion> = tracking_input
+        .mouse
+        .then(|| mouse_motion.read().cloned().collect())
+        .unwrap_or_default();
 
-    // `wants_pointer_input` sometimes returns `false` after clicking or holding a button over a widget,
-    // so `is_pointer_over_area` is also needed.
-    #[cfg(feature = "egui")]
-    let (mouse_buttons, mouse_wheel) = if maybe_egui.iter_mut().any(|(_, mut ctx)| {
-        ctx.get_mut().is_pointer_over_area() || ctx.get_mut().wants_pointer_input()
-    }) {
-        (None, None)
-    } else {
-        (mouse_buttons, mouse_wheel)
-    };
+    // Ensure the conflicted inputs will be tracked next time
+    tracking_input.tick();
 
     let resources = input_map
         .zip(action_state)